% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_sums_means.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate Student Level to School/District/State}
\usage{
aggregate(df, op_list = list(), ...)
}
\arguments{
\item{df}{Dataframe. Used first to be friendly to the \%>\% operator.}

\item{op_list}{A named list where the name is the operation and the value is a character vector
of column names to apply the operation to}

\item{...}{The columns to aggregate at (e.g. system, system_name, school, school_name)}
}
\description{
This function takes a student level dataframe and aggregates it to the school/district/state level.
Provides sums, means, medians, standard deviation, or variance for desired columns.
}
\examples{
aggregate(student_level_df,op_list = list(sum = c('enrolled', 'tested', 'valid tests')),
system, school, subgroup)
}
\keyword{aggregate,}
\keyword{collapse}
\keyword{group,}
