% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amo_target.R
\name{amo_target}
\alias{amo_target}
\title{Calculate AMO Targets}
\usage{
amo_target(n, pct, double = FALSE)
}
\arguments{
\item{n}{A count (valid tests, enrolled, etc.).}

\item{pct}{A percentage (Percent On Track/Mastered, Chronic Absenteeism, etc.).}

\item{double}{Calculate double AMO target? Defaults to FALSE.}
}
\description{
Family of functions for calcuating AMO targets. Includes functions for
AMO reduction targets and double AMO targets.
}
\examples{

}
