% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_pct_cols.R
\name{bind_pct_cols}
\alias{bind_pct_cols}
\title{Calculates Percentages and Adds them to the end of the data frame}
\usage{
bind_pct_cols(df, numerator_cols, denom_col)
}
\arguments{
\item{df}{Dataframe. Used first to be friendly to the \%>\% operator.}

\item{numerator_cols}{A character vector of columns to be used in the numerator.}

\item{denom_col}{A single string of the column name to be used in the denominator.}
}
\value{
Returns a dataframe with the pct columns at the end. The columns will be named
the original column name with a 'pct_' prefix
}
\description{
This function takes a dataframe and a list of columns to find a percentage for, along with a single column to use
in the denominator. It returns a dataframe
}
\examples{
bind_pct_cols(system_level_df, c('on_track', 'mastered'), 'valid_tests')
}
\keyword{calculate}
\keyword{percent,}
\keyword{percentage,}
