% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_upper_bound_ci.R
\name{ci_lower_bound}
\alias{ci_lower_bound}
\title{Calculate Upper/Lower Bound Confidence Intervals}
\usage{
ci_lower_bound(n, pct)
}
\arguments{
\item{n}{A count (valid tests, enrolled, etc.).}

\item{pct}{A percentage (Percent On Track/Mastered, Chronic Absenteeism, etc.).}
}
\description{
Calculate Upper/Lower Bound Confidence Intervals
}
\examples{

}
