% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_distinct.R
\name{count_distinct}
\alias{count_distinct}
\title{Count distinct combinations of a set of variables}
\usage{
count_distinct(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{An unquoted set of variables, passed to dplyr::select.}
}
\description{
Count distinct combinations of a set of variables
}
\examples{

count_distinct(iris, Species)
}
