% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_if.R
\name{count_if}
\alias{count_if}
\title{Count Observations in a Data Frame Meeting a Condition}
\usage{
count_if(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{A logical condition, passed to dplyr::filter. Multiple conditions are combined with &.}
}
\description{
Count Observations in a Data Frame Meeting a Condition
}
\examples{

count_if(iris, Species == "setosa")
}
