% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{not_na}
\alias{not_na}
\title{Functions to check (non-)missingness.}
\usage{
not_na(x)
}
\arguments{
\item{x}{A vector.}

\item{df}{A data frame.}
}
\description{
Functions to check (non-)missingness.
}
