% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round5.R
\name{round5}
\alias{round5}
\title{Round Numbers}
\usage{
round5(x, digits = 0)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{digits}{Number of decimal places. Defaults to 0 (round to the whole number).}
}
\description{
Does not handle rounding of .5 in the usual R way.
}
\examples{

round(.5 + -2:4) # -2  0  0  2  2  4  4
round5(.5 + -2:4) # -2  -1  1  2  3  4  5

}
