% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/View_at_if.R
\name{View_at}
\alias{View_at}
\title{View a subset of a data frame.}
\usage{
View_at(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{An unquoted set of variables, passed to dplyr::select.}
}
\description{
View a subset of a data frame.
}
\examples{

View_at(iris, Species, Sepal.Length, Petal.Length)

}
