% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/View_at_if.R
\name{View_if}
\alias{View_if}
\title{View a subset of rows from a data frame.}
\usage{
View_if(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{A logical condition, passed to dplyr::filter. Multiple conditions are combined with &.}
}
\description{
View a subset of rows from a data frame.
}
\examples{

View_if(iris, Species == "setosa")
}
