% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amo_target.R
\name{amo_target}
\alias{amo_target}
\title{Calculate AMO Targets}
\usage{
amo_target(n, pct, double = FALSE, n_minimum = 30)
}
\arguments{
\item{n}{A count (valid tests, enrolled, etc.). Targets are only calculated if
count is 30 or greater.}

\item{pct}{A percentage (Percent On Track/Mastered, Chronic Absenteeism, etc.).}

\item{double}{Calculate double AMO target? Defaults to FALSE.}

\item{n_minimum}{Minimum N count to calculate AMO target. Defaults to 30.}
}
\description{
Family of functions for calcuating AMO targets. Includes functions for
AMO reduction targets and double AMO targets.
}
\examples{

amo_target(n = 30, pct = 36)

# Equivalent
amo_reduction(n = 50, pct = 20.7, double = TRUE)
amo_reduction_double(n = 50, pct = 20.7)

}
