% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/View_at_if.R
\name{View_at}
\alias{View_at}
\title{View a subset of a data frame.}
\usage{
View_at(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{A logical condition (View_if) or an unquoted set of variables (View_at).}
}
\description{
View a subset of a data frame.
}
\examples{

df <- data.frame(
    x = sample(LETTERS, 1000, replace = TRUE),
    y = rnorm(1000),
    z = rnorm(1000)
)

View_if(df, x == "A")

View_at(df, x, z)

}
