% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/View_at_if.R
\name{View_if}
\alias{View_if}
\title{View a subset of rows from a data frame.}
\usage{
View_if(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{A logical condition.}
}
\description{
View a subset of rows from a data frame.
}
\examples{

df <- data.frame(
    x = sample(LETTERS, 1000, replace = TRUE),
    y = rnorm(1000),
    z = rnorm(1000)
)

View_if(df, x == "A")

}
